apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.9
httpx==0.27.2
Pillow==11.0.0
numpy==2.1.3
# models
transformers==4.45.2
torch==2.4.0
torchvision==0.19.0
open-clip-torch==2.26.1
# storage
SQLAlchemy==2.0.36
psycopg[binary,pool]==3.2.3
qdrant-client==1.11.3
pgvector==0.3.3
# metrics & tracing
prometheus-client==0.21.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp==1.27.0
python-dotenv==1.0.1
